refactor(fuzz): split fixture and invariants out of main.rs
main.rs keeps only the #[invariant_test] entry points; the world and
its actions live in fixture.rs and reusable checks in invariants.rs, so
scoped tests (lifecycle, delegations, plan admin, full) can share both
as the action set grows.
